moved cancelAirIncrement to LivingBreathEvent
