+++ fabric/src/main/java/draylar/identity/fabric/mixin/LivingEntityMixin.java
@@ -1,0 +1,40 @@
+package draylar.identity.mixin;
+
+import draylar.identity.Identity;
+import draylar.identity.api.PlayerIdentity;
+import net.minecraft.entity.*;
+import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+@Mixin(LivingEntity.class)
+public abstract class LivingEntityMixin extends Entity {
+
+    @Shadow
+    protected abstract int getNextAirOnLand(int air);
+
+    protected LivingEntityMixin(EntityType<?> type, World world) {
+        super(type, world);
+    }
+
+    @Redirect(
+            method = "baseTick",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/LivingEntity;setAir(I)V", ordinal = 2)
+    )
+    private void cancelAirIncrement(LivingEntity livingEntity, int air) {
+        // Aquatic creatures should not regenerate breath on land
+        if ((Object) this instanceof PlayerEntity player) {
+            LivingEntity identity = PlayerIdentity.getIdentity(player);
+
+            if (identity != null) {
+                if (Identity.isAquatic(identity)) {
+                    return;
+                }
+            }
+        }
+
+        this.setAir(this.getNextAirOnLand(this.getAir()));
+    }
+}
