+++ gradle.properties
@@ -10,7 +10,7 @@ maven_group=dev.draylar
 fabric_loader_version=0.14.21
 fabric_api_version=0.84.0+1.20.1
 yarn_mappings=1.20.1+build.8
-forge_version=47.0.43
+forge_version=47.1.43
 
 # Dependency Versions
 architectury_version=9.0.8
